# src/main/resources/static/openapi.yml
openapi: 3.0.3
info:
  title: API Rest - Backend - Arena Mafia
  version: v1
  description: |
    Backend do projeto para a Arena Mafia, desenvolvido como parte do curso de
    Análise e Desenvolvimento de Sistemas do Instituto Federal de Santa Catarina.
  termsOfService: https://github.com/LucasRibasCardoso/backend-arena-mafia
  license:
    name: Apache 2.0
    url: https://github.com/LucasRibasCardoso
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Token de autenticação JWT obtido no login."
  schemas:
    FieldErrorResponseDto:
      type: object
      properties:
        fieldName: {
          type: string,
          description: "Nome do campo que falhou na validação.",
          example: "username"
        }
        message: {
          type: string,
          description: "Mensagem de erro para o campo.",
          example: "O usuário é obrigatório"
        }
      required: [ fieldName, message ]
    ErrorResponseDto:
      type: object
      properties:
        timestamp: {
          type: string,
          format: date-time,
          example: "2025-08-15T13:55:00Z"
        }
        status: {
          type: integer,
          example: 400
        }
        message: {
          type: string,
          example: "Requisição inválida."
        }
        path: {
          type: string,
          example: "/api/some/endpoint"
        }
        fieldErrors:
          type: array
          description: "Lista de erros de validação de campos, se houver."
          items: { $ref: "#/components/schemas/FieldErrorResponseDto" }
          nullable: true
      required: [ timestamp, status, message, path, fieldErrors ]
...
    TokenResponseDto:
      type: object
      properties:
        phone: {
          type: string,
          description: "Telefone do usuário autenticado.",
          example: "+5547912345678"
        }
        username: {
          type: string,
          description: "Nome de usuário autenticado.",
          example: "usuario123"
        }
        fullName: {
          type: string,
          description: "Nome completo do usuário autenticado.",
          example: "João da Silva"
        }
        role: {
          type: string,
          description: "Permissão do usuário autenticado.",
          example: "ROLE_USER"
        }
        accessToken: {
          type: string,
          description: "JWT de acesso.",
          example: "eyJhbGci..."
        }
...
    ValidateOtpRequestDto:
      type: object
      description: "Dados para verificação da conta via código OTP."
      required: [ phone, code ]
      properties:
        phone:
          type: string
          description: "Telefone de usuário da conta a ser verificada."
          example: "+5547912345678"
        code:
          type: string
          description: "Código OTP de 6 dígitos recebido por SMS."
          example: "123456"
    ResendCodeRequestDto:
      type: object
      description: "Telefone do usuário que está solicitando o reenvio do código OTP."
      required: [ phone ]
      properties:
        phone:
          type: string
          description: "Telefone de usuário da conta para a qual o código deve ser reenviado."
          example: "+5547912345678"
...
  headers:
    Cookie:
      description: "Cookie HttpOnly contendo o token de refresh. O navegador o armazena automaticamente e o anexa às requisições subsequentes."
      schema:
        type: string
        example: "refreshToken=asdasda-asdasd-asdasd-asdasd..."
...
  responses:
    ErrorResponseInternalServerError:
      description: Erro interno do servidor.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseDto"
          example:
            timestamp: "2025-08-15T13:55:00Z"
            status: 500
            message: "Erro interno do servidor. Por favor, tente novamente mais tarde."
            path: "/api/some/endpoint"
...
    TokenResponseDto:
      description: "Usuário autenticado com sucesso."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/TokenResponseDto"
...
paths:
...
  /api/auth/refresh-token:
    post:
      operationId: "refreshToken"
      tags:
        - "Autenticação"
      summary: "Renovar o token de acesso"
      description: "Gera um novo token de acesso (JWT) a partir de um refresh token válido. O refresh token deve ser enviado em um cookie HttpOnly chamado 'refreshToken'."
      parameters:
        - in: "cookie"
          name: "refreshToken"
          required: true
          description: "O refresh token recebido durante o login. O navegador anexa este cookie automaticamente."
          schema:
            type: string
            example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
      responses:
        "200":
          description: "Token de acesso renovado com sucesso. Um novo access token é retornado no corpo da resposta e um novo refresh token é enviado no cabeçalho 'Set-Cookie'."
          headers:
            Cookie:
              $ref: "#/components/headers/Cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponseDto"
        "400":
          description: "Requisição inválida. Ocorre se o formato do 'refreshToken' for inválido (não é um UUID válido)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              example:
                timestamp: "2025-08-15T14:21:00Z"
                status: 400
                message: "Formato inválido para o refresh token."
                path: "/api/auth/refresh-token"
        "401":
          description: "Não autorizado. Ocorre se o 'refreshToken' estiver ausente, expirado ou não for encontrado no sistema."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                MissingToken:
                  summary: "Refresh token ausente"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 401
                    message: "Sua sessão expirou. Por favor, faça login novamente."
                    path: "/api/auth/refresh-token"
                TokenNotFound:
                  summary: "Refresh token não encontrado"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 401
                    message: "Refresh token não encontrado."
                    path: "/api/auth/refresh-token"
        "409":
          description: "Conflito de estado. Ocorre se a conta não estiver ativada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                PendingVerification:
                  summary: "Conta pendente de verificação"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Você precisa ativar sua conta. Por favor, termine o processo de cadastro."
                    path: "/api/auth/refresh-token"
                Locked:
                  summary: "Conta bloqueada"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está bloqueada. Por favor, contate o suporte."
                    path: "/api/auth/refresh-token"
                Disabled:
                  summary: "Conta desativada"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 409
                    message: "Atenção: Sua conta está desativada."
                    path: "/api/auth/refresh-token"
        "500":
          $ref: "#/components/responses/ErrorResponseInternalServerError"
...
  /api/auth/verify-account:
    post:
      operationId: "verifyAccount"
      tags:
        - "Autenticação"
      summary: "Verificar e ativar uma nova conta"
      description: "Valida o código de verificação (OTP) enviado por SMS para uma conta com status 'PENDING_VERIFICATION'. Se o código for válido, a conta é ativada (status 'ACTIVE'), e o usuário é autenticado, recebendo um token de acesso e um cookie 'refreshToken'."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateOtpRequestDto"
      responses:
        "200":
          description: "Conta ativada e usuário autenticado com sucesso."
          headers:
            Cookie:
              $ref: "#/components/headers/Cookie"
          content:
            application/json:
              schema:
                $ref: "#/components/responses/TokenResponseDto"
        "400":
          description: "Requisição inválida. Ocorre se os dados forem inválidos (ex: telefone, código em branco) ou se o código OTP for incorreto ou expirado."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              examples:
                MissingCode:
                  summary: "Código de verificação ausente"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Erro de validação. Verifique os campos informados."
                    path: "/api/auth/verify-account"
                    fieldErrors:
                      - fieldName: "code"
                        message: "O código de verificação é obrigatório."
                InvalidPhone:
                  summary: "Número de telefone inválido"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Número de telefone inválido. Verifique o DDD e a quantidade de dígitos."
                    path: "/api/auth/verify-account"
                InvalidCode:
                  summary: "Código OTP inválido ou expirado"
                  value:
                    timestamp: "2025-08-15T13:55:00Z"
                    status: 400
                    message: "Código de verificação inválido ou expirado."
                    path: "/api/auth/verify-account"
        "404":
          description: "Não encontrado. Ocorre se não houver um usuário com o telefone fornecido e status pendente."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              example:
                timestamp: "2025-08-15T13:55:00Z"
                status: 404
                message: "Usuário não encontrado. Retorne ao início do cadastro para criar uma nova conta."
                path: "/api/auth/verify-account"
        "409":
          description: "Conflito de estado. Ocorre ao tentar verificar uma conta que já se encontra em um estado 'ACTIVE"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
              example:
                timestamp: "2025-08-15T13:55:00Z"
                status: 409
                message: "Atenção: A conta já está ativada. Você pode fazer login."
                path: "/api/auth/verify-account"
        "500":
          $ref: "#/components/responses/ErrorResponseInternalServerError"
  /api/auth/resend-code:
    post:
      operationId: "resendVerificationCode"
      tags:
        - "Autenticação"
      summary: "Reenviar código de verificação"
      description: "Gera e reenvia um novo código de verificação (OTP) para o telefone de um usuário com conta em estado 'PENDING_VERIFICATION'. Por segurança, este endpoint sempre retorna sucesso, mesmo que o telefone não esteja cadastrado, para evitar a enumeração de usuários."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResendCodeRequestDto"
        responses:
          "204":
            description: "Sucesso. A solicitação foi validada e um novo código de verificação foi enviado para o número de telefone associado à conta. O cliente deve agora permitir que o usuário insira o novo código."
          "400":
            description: "Requisição inválida. Ocorre se o formato do telefone for inválido."
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponseDto"
                examples:
                  MissingPhone:
                    summary: "Número de telefone não informado"
                    value:
                      timestamp: "2025-08-15T13:55:00Z"
                      status: 400
                      message: "Erro de validação. Verifique os campos informados."
                      path: "/api/auth/resend-code"
                      fieldErros:
                        fieldName: "phone"
                        message: "Por favor, informe seu telefone."
                  InvalidPhone:
                    summary: "Número de telefone inválido"
                    value:
                      timestamp: "2025-08-15T13:55:00Z"
                      status: 400
                      message: "Número de telefone inválido. Verifique o DDD e a quantidade de dígitos."
                      path: "/api/auth/resend-code"
          "409":
            description: "Conflito de estado. Ocorre ao tentar reenviar um código para uma conta que não está com status 'PENDING_VERIFICATION'."
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponseDto"
                example:
                  timestamp: "2025-08-15T13:55:00Z"
                  status: 409
                  message: "Atenção: Só é possível reenviar o código para contas pendentes de verificação."
                  path: "/api/auth/resend-code"
          "500":
            $ref: "#/components/responses/ErrorResponseInternalServerError"
...